Remove invalid test for `Array.prototype.sort`
Neither ES5 nor ES2015 specify that an exception should be thrown when
`Array.prototype.sort` is invoked with a non-function, non-undefined
value. The most relevant specification text reads:

> If comparefn is not undefined and is not a consistent comparison
> function for the elements of this array (see below), the sort order is
> implementation-defined.

So while an exception is a logical response to invalid input, the
specification does not require this.

Remove the test that enforces this requirement.